Extract presentation protocol FSMs
