+++ aries_vcx/src/protocols/mod.rs
@@ -1,4 +1,5 @@
 pub mod connection;
 pub mod issuance;
+pub mod proof_presentation;
 // pub mod proof_presentation;
 // pub mod out_of_band;
