+++ aries_vcx/src/protocols/proof_presentation/mod.rs
@@ -1,0 +1,2 @@
+pub mod prover;
+pub mod verifier;
