+++ aries_vcx/src/protocols/proof_presentation/prover/states/presentation_prepared.rs
@@ -1,5 +1,5 @@
-use crate::handlers::proof_presentation::prover::states::finished::FinishedState;
-use crate::handlers::proof_presentation::prover::states::presentation_sent::PresentationSentState;
+use crate::protocols::proof_presentation::prover::states::finished::FinishedState;
+use crate::protocols::proof_presentation::prover::states::presentation_sent::PresentationSentState;
 use crate::messages::proof_presentation::presentation::Presentation;
 use crate::messages::proof_presentation::presentation_request::PresentationRequest;
 use crate::messages::status::Status;
