+++ aries_vcx/src/protocols/proof_presentation/verifier/states/finished.rs
@@ -1,8 +1,8 @@
-use crate::handlers::proof_presentation::verifier::state_machine::RevocationStatus;
+use crate::messages::error::ProblemReport;
 use crate::messages::proof_presentation::presentation::Presentation;
 use crate::messages::proof_presentation::presentation_request::PresentationRequest;
 use crate::messages::status::Status;
-use crate::messages::error::ProblemReport;
+use crate::protocols::proof_presentation::verifier::state_machine::RevocationStatus;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct FinishedState {
