+++ aries_vcx/src/protocols/proof_presentation/verifier/states/presentation_request_set.rs
@@ -1,5 +1,5 @@
-use crate::handlers::proof_presentation::verifier::states::presentation_request_sent::PresentationRequestSentState;
 use crate::messages::proof_presentation::presentation_request::PresentationRequest;
+use crate::protocols::proof_presentation::verifier::states::presentation_request_sent::PresentationRequestSentState;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
 pub struct PresentationRequestSetState {
